requirements.txt
openai>=1.0.0
pandas>=1.5.0
tqdm>=4.65.0
python-dotenv>=0.19.0
tenacity>=8.2.0
click>=8.1.0
openpyxl>=3.1.0
python-json-logger>=2.0.0